backend/requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
isort==5.10.1
mypy==0.971
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
PyJWT==2.6.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0